Fixup uuid stubbing in test_stack_user.py
As noted in https://review.openstack.org/#/c/71930/13  there is an issue
with using the utils.UUIDStub method to stub the resource ID, because
any events stored also get the same uuid, which can cause sqlalchemy
warnings to be output.  Instead store the resource before setting up the
keystoneclient stub so we know the expected ID to use.

Change-Id: Ia456425384d52b3983b8d1d387f56e57e8b559ee